+++ backend/backend/src/main/java/ltd/hengpro/backend/vo/SiteStatisticsVo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class SiteStatisticsVo {
 
-    private Integer userNum;
+    private Long userNum;
 
-    private Integer catNum;
+    private Long catNum;
 }
